test: move unit test to their own package
