+++ pkg/handlers/generic/mutation/containerdrestart/inject_test.go
@@ -1,30 +1,33 @@
 // Copyright 2023 D2iQ, Inc. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package tests
+package containerdrestart
 
 import (
 	"testing"
 
+	. "github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
 	runtimehooksv1 "sigs.k8s.io/cluster-api/exp/runtime/hooks/api/v1alpha1"
 
 	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/common/pkg/capi/clustertopology/handlers/mutation"
 	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/common/pkg/testutils/capitest"
 	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/common/pkg/testutils/capitest/request"
-	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/pkg/handlers/generic/mutation/containerdrestart"
 )
 
-func TestGeneratePatches(
-	t *testing.T,
-	generatorFunc func() mutation.GeneratePatches,
-) {
-	t.Helper()
+func TestContainerdRestartPatch(t *testing.T) {
+	gomega.RegisterFailHandler(Fail)
+	RunSpecs(t, "Containerd restart mutator suite")
+}
 
-	capitest.ValidateGeneratePatches(
-		t,
-		generatorFunc,
-		capitest.PatchTestDef{
+var _ = Describe("Generate Containerd restart patches", func() {
+	// only add aws region patch
+	patchGenerator := func() mutation.GeneratePatches {
+		return mutation.NewMetaGeneratePatchesHandler("", NewPatch()).(mutation.GeneratePatches)
+	}
+
+	testDefs := []capitest.PatchTestDef{
+		{
 			Name:        "restart script and command added to control plane kubeadm config spec",
 			RequestItem: request.NewKubeadmControlPlaneTemplateRequestItem(""),
 			ExpectedPatchMatchers: []capitest.JSONPatchMatcher{
@@ -33,7 +36,7 @@ func TestGeneratePatches(
 					Path:      "/spec/template/spec/kubeadmConfigSpec/files",
 					ValueMatcher: gomega.ContainElements(
 						gomega.HaveKeyWithValue(
-							"path", containerdrestart.ContainerdRestartScriptOnRemote,
+							"path", ContainerdRestartScriptOnRemote,
 						),
 					),
 				},
@@ -41,12 +44,12 @@ func TestGeneratePatches(
 					Operation: "add",
 					Path:      "/spec/template/spec/kubeadmConfigSpec/preKubeadmCommands",
 					ValueMatcher: gomega.ContainElements(
-						containerdrestart.ContainerdRestartScriptOnRemoteCommand,
+						ContainerdRestartScriptOnRemoteCommand,
 					),
 				},
 			},
 		},
-		capitest.PatchTestDef{
+		{
 			Name: "restart script and command added to worker node kubeadm config template",
 			Vars: []runtimehooksv1.Variable{
 				capitest.VariableWithValue(
@@ -65,7 +68,7 @@ func TestGeneratePatches(
 					Path:      "/spec/template/spec/files",
 					ValueMatcher: gomega.ContainElements(
 						gomega.HaveKeyWithValue(
-							"path", containerdrestart.ContainerdRestartScriptOnRemote,
+							"path", ContainerdRestartScriptOnRemote,
 						),
 					),
 				},
@@ -73,10 +76,22 @@ func TestGeneratePatches(
 					Operation: "add",
 					Path:      "/spec/template/spec/preKubeadmCommands",
 					ValueMatcher: gomega.ContainElements(
-						containerdrestart.ContainerdRestartScriptOnRemoteCommand,
+						ContainerdRestartScriptOnRemoteCommand,
 					),
 				},
 			},
 		},
-	)
-}
+	}
+
+	// create test node for each case
+	for testIdx := range testDefs {
+		tt := testDefs[testIdx]
+		It(tt.Name, func() {
+			capitest.AssertGeneratePatches(
+				GinkgoT(),
+				patchGenerator,
+				&tt,
+			)
+		})
+	}
+})
